Restructure the personal calculator files into a separate folder Fixes #13
