+++ personal_calculator/calculator.py
@@ -1,7 +1,6 @@
 from policyengine_us import Simulation
 from policyengine_core.reforms import Reform
-from reforms import PolicyReforms
-
+from personal_calculator.reforms import PolicyReforms
 
 def calculate_impacts(situation, reform_params_dict):
     """Calculate the impacts of multiple reform scenarios"""
