+++ personal_calculator/inputs.py
