+++ personal_calculator/situation.py
