+++ personal_calculator/table.py
